♻️ Refactor: #349 Enforce ui barrel imports via exports map
Fix the four deep ui/src imports; globals.css kept as an explicit subpath.
