+++ packages/ui/package.json
@@ -3,6 +3,13 @@
   "version": "0.0.0",
   "main": "./index.ts",
   "types": "./index.ts",
+  "exports": {
+    ".": {
+      "types": "./index.ts",
+      "default": "./index.ts"
+    },
+    "./src/globals.css": "./src/globals.css"
+  },
   "license": "MIT",
   "scripts": {
     "lint": "biome check .",
